Move "Controller" to customer namespace, revert the original one
